+++ docs/history.rst
@@ -8,7 +8,7 @@ History
 Version 0.5.3 - Unreleased
 --------------------------
 
-* Removed path prefix ``articles/`` from detail view since it is useless;
+* Removed path prefix ``articles/`` from detail view URL since it is useless;
 * Added new setting ``LOTUS_CRUMBS_TITLES`` so crumb title for views can be customized
   from settings. However this does not apply for detail views which directly use their
   object title as the crumb title;
@@ -19,6 +19,8 @@ Version 0.5.3 - Unreleased
 * Added ``seo_title``, ``lead``, ``introduction``, ``content`` to seachable fields in
   Article admin list;
 * Added ``lead``, ``description`` to seachable fields in Category admin list;
+* Added new filter to Article admin list to filter on published or unpublished items;
+* Moved admin filters from ``lotus.admin.translated`` to ``lotus.admin_filters``;
 
 
 Version 0.5.2.1 - 2023/06/03